Wrangled the GoL into an application subclass: - TODO: clean out and abstract useful parts (e.g camera, shaders etc)
