+++ GL.h
@@ -1,0 +1,7 @@
+#pragma once
+#include <GL/glew.h>
+#include <glm/glm.hpp>
+#include <glm/gtc/type_ptr.hpp>
+#include <glm/gtc/matrix_transform.hpp>
+
+#include "Application.h"
